RST docs support

+++ bot/docs.py
@@ -23,9 +23,13 @@ def generate_docs():
         os.remove("./docs/plugins/" + file)
     for plugin in os.listdir('./plugins/'):
         if plugin[0] != '_':
-            if os.path.isfile('./plugins/' + plugin + "/docs/user_documentation.md"):
-                copyfile('./plugins/' + plugin + "/docs/user_documentation.md", './docs/plugins/' + plugin + ".md")
-                docs.write("    plugins/" + plugin + ".md\n")
+            if os.path.isfile('./plugins/' + plugin + "/docs/user_documentation.rst"):
+                copyfile('./plugins/' + plugin + "/docs/user_documentation.md", './docs/plugins/' + plugin + ".rst")
+                docs.write("    plugins/" + plugin + ".rst\n")
+            else:
+                if os.path.isfile('./plugins/' + plugin + "/docs/user_documentation.md"):
+                    copyfile('./plugins/' + plugin + "/docs/user_documentation.md", './docs/plugins/' + plugin + ".md")
+                    docs.write("    plugins/" + plugin + ".md\n")
 
 
     if os.listdir('./docs/create_plugin') != []:
